refactored to move solr common objects to orcid-utils
