+++ orcid-message-listener/src/main/resources/message-listener.properties
@@ -1,0 +1,27 @@
+#
+# =============================================================================
+#
+# ORCID (R) Open Source
+# http://orcid.org
+#
+# Copyright (c) 2012-2014 ORCID, Inc.
+# Licensed under an MIT-Style License (MIT)
+# http://orcid.org/open-source-license
+#
+# This copyright and license information (including a link to the full license)
+# shall be included in its entirety in all copies or substantial portion of
+# the software.
+#
+# =============================================================================
+#
+
+org.orcid.persistence.solr.url=http://localhost:8080/orcid-solr-web
+org.orcid.persistence.solr.read.only.url=http://localhost:8080/orcid-solr-web
+org.orcid.persistence.solr.socketTimeout=2000
+org.orcid.persistence.solr.connectionTimeout=1000
+org.orcid.persistence.solr.allowCompression=true
+org.orcid.persistence.solr.maxRetries=1
+org.orcid.persistence.solr.defaultMaxConnectionsPerHost=100
+org.orcid.persistence.solr.maxTotalConnections=100
+
+org.orcid.persistence.messaging.brokerURL=tcp://localhost:61616?jms.useAsyncSend=true
