+++ orcid-message-listener/src/main/resources/orcid-message-listener-web-context.xml
@@ -26,38 +26,62 @@
         http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.1.xsd
         http://www.springframework.org/schema/cache http://www.springframework.org/schema/cache/spring-cache.xsd
         http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop-3.1.xsd 
-        http://www.springframework.org/schema/jms http://www.springframework.org/schema/jms/spring-jms-4.1.xsd" >
+        http://www.springframework.org/schema/jms http://www.springframework.org/schema/jms/spring-jms-4.1.xsd">
 
-    <context:component-scan
-		base-package="org.orcid.listener" />
-		
-    	<!-- JMS Messaging -->
+	<context:component-scan base-package="org.orcid.listener" />
+
+	<!-- ${org.orcid.persistence.properties} gets replaced during the Maven 
+		build, but other properties are escaped from Maven build replacement by preceding 
+		with \ and so are handled by Spring. -->
+	<context:property-placeholder location="\${org.orcid.config.file:}"
+		ignore-resource-not-found="true" ignore-unresolvable="true" />
+	<context:property-placeholder
+		location="classpath:${org.orcid.message-listener.properties}"
+		ignore-resource-not-found="true" ignore-unresolvable="true" />
+
+	<!--SOLR -->
+	<!-- Confusingly, the client is actually called a server -->
+	<bean id="solrServer" class="org.apache.solr.client.solrj.impl.HttpSolrServer">
+		<constructor-arg value="\${org.orcid.persistence.solr.url}" />
+		<property name="soTimeout" value="\${org.orcid.persistence.solr.socketTimeout}" />
+		<property name="connectionTimeout" value="\${org.orcid.persistence.solr.connectionTimeout}" />
+		<property name="allowCompression" value="\${org.orcid.persistence.solr.allowCompression}" />
+		<property name="maxRetries" value="\${org.orcid.persistence.solr.maxRetries}" />
+		<property name="defaultMaxConnectionsPerHost" value="\${org.orcid.persistence.solr.maxTotalConnections}" />
+		<property name="maxTotalConnections" value="\${org.orcid.persistence.solr.defaultMaxConnectionsPerHost}" />
+		<property name="parser" ref="responseParser" />
+	</bean>
+
+	<bean name="responseParser" class="org.apache.solr.client.solrj.impl.XMLResponseParser"></bean>
+
+	<!-- JMS Messaging -->
 	<!-- enable the configuration of jms on annotations -->
-	<jms:annotation-driven/>
+	<jms:annotation-driven />
 	<bean id="jmsConnectionFactory" class="org.apache.activemq.ActiveMQConnectionFactory">
-		<property name="brokerURL" value="tcp://0.0.0.0:61616"/>
+		<property name="brokerURL" value="${org.orcid.persistence.messaging.brokerURL}" />
 	</bean>
-	<bean id="cachingJmsConnectionFactory" class="org.springframework.jms.connection.CachingConnectionFactory">
-    	<property name="targetConnectionFactory" ref="jmsConnectionFactory"/>
+	<bean id="cachingJmsConnectionFactory"
+		class="org.springframework.jms.connection.CachingConnectionFactory">
+		<property name="targetConnectionFactory" ref="jmsConnectionFactory" />
 	</bean>
-	
+
 	<!-- so we can listen to queues -->
 	<bean id="jmsListenerContainerFactory" class="org.springframework.jms.config.DefaultJmsListenerContainerFactory">
-	    <property name="connectionFactory" ref="jmsConnectionFactory" />
-	    <property name="concurrency" value="1"/>
-	    <!-- property name="concurrency" value="1-10"/ -->
+		<property name="connectionFactory" ref="jmsConnectionFactory" />
+		<property name="concurrency" value="1" />
+		<!-- property name="concurrency" value="1-10"/ -->
 	</bean>
-	
+
 	<!-- Default Destination Queue Definition -->
 	<bean id="defaultJmsDestination" class="org.apache.activemq.command.ActiveMQQueue">
-	  <!-- name of the queue -->
-	  <constructor-arg index="0" value="test" />
+		<!-- name of the queue -->
+		<constructor-arg index="0" value="test" />
 	</bean>
-	 
+
 	<!-- JmsTemplate Definition so we can send to queues -->
 	<bean id="jmsTemplate" class="org.springframework.jms.core.JmsTemplate">
-	  <property name="connectionFactory" ref="jmsConnectionFactory" />
-	  <property name="defaultDestination" ref="defaultJmsDestination" />
+		<property name="connectionFactory" ref="jmsConnectionFactory" />
+		<property name="defaultDestination" ref="defaultJmsDestination" />
 	</bean>
-	
+
 </beans>
