+++ orcid-utils/pom.xml
@@ -43,6 +43,11 @@
 		    <artifactId>guava</artifactId>
 		    <version>18.0</version>
         </dependency>
+        
+        <dependency>
+            <groupId>org.apache.solr</groupId>
+            <artifactId>solr-solrj</artifactId>
+        </dependency>
 
         <!-- Test dependencies -->
