+++ orcid-utils/src/main/java/org/orcid/utils/solr/entities/OrcidSolrResult.java
@@ -14,7 +14,7 @@
  *
  * =============================================================================
  */
-package org.orcid.persistence.solr.entities;
+package org.orcid.utils.solr.entities;
 
 import java.util.Collection;
